change signature of part 1 funcs to be == to the man

+++ ft_bzero.c
@@ -6,15 +6,15 @@
 /*   By: jbonnin <jbonnin@42student.fr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/09 15:31:36 by jbonnin           #+#    #+#             */
-/*   Updated: 2023/01/14 20:04:23 by jbonnin          ###   ########.fr       */
+/*   Updated: 2023/01/17 19:59:33 by jbonnin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdio.h>
+#include "libft.h"
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*res;
 
 	i = 0;
